+++ app/models/caffeinate/campaign_subscription.rb
@@ -58,7 +58,7 @@ module Caffeinate
 
     after_create :create_mailings!
 
-    after_commit :on_complete, if: :completed?
+    after_commit :on_complete, if: :completed?, unless: :destroyed?
 
     # Add (new) drips to a `CampaignSubscriber`.
     #
@@ -99,6 +99,7 @@ module Caffeinate
     # Updates `ended_at` and runs `on_complete` callbacks
     def end!(reason = ::Caffeinate.config.default_ended_reason)
       raise ::Caffeinate::InvalidState, 'CampaignSubscription is already unsubscribed.' if unsubscribed?
+      return true if ended?
 
       update!(ended_at: ::Caffeinate.config.time_now, ended_reason: reason)
 
@@ -109,6 +110,7 @@ module Caffeinate
     # Updates `ended_at` and runs `on_complete` callbacks
     def end(reason = ::Caffeinate.config.default_ended_reason)
       return false if unsubscribed?
+      return true if ended?
 
       result = update(ended_at: ::Caffeinate.config.time_now, ended_reason: reason)