+++ spec/rails_helper.rb
@@ -65,12 +65,22 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
 end
 
+puts "bundle exec rspec --seed #{RSpec.configuration.seed}"
+
 ActiveJob::Base.queue_adapter = :test
 Rails.application.routes.default_url_options[:host] = 'localhost:3000'
 
 class BaseTestDripper < ::Caffeinate::Dripper::Base; end
 class ArgumentMailer < ActionMailer::Base
   def hello(_mailing)
+    mail(to: 'argument_mailer@example.com', from: 'argument_mailer@example.com', subject: 'Argument Mailer') do |format|
+      format.text { render plain: 'ArgumentMailer' }
+    end
+  end
+end
+
+class AnyMailer < ActionMailer::Base
+  def welcome(_mailing)
     mail(to: 'argument_mailer@example.com', from: 'argument_mailer@example.com', subject: 'Argument Mailer') do |format|
       format.text { render plain: 'ArgumentMailer' }
     end